fix initial history handling
When there's no history, we have to put in a false #shellswain init entry
in order for the fc command to work. This commit fixes a bug in when it was
added in, and adds a mechanism for popping it back out safely to avoid polluting
the history over time.