refine time tracking for multi-plugin environments
Previously, record_start and record_end were wired up as before/after
events. This worked fine, but bashup.events runs the events in
the order they are registered. This means start time and end time were
both collected before plugin event handlers run.

Basically, any plugin-registered before_command listeners would get
counted as part of the command duration, while any after_command
listeners would be excluded from the duration.

This change inlines the contents of _record_start because it was a
lukewarm fit for the two contexts it was called in already. Inlining
makes it easy to optimize where the start timestamps are collected.

Potential gotcha regarding the availability of start_timestamp
within before_command listeners. Since it was previously set before
these ran, I decided to go ahead and collect it twice:
- before emitting the event, so that _something_ is available if
  there's a clear need for it
- after emitting the event, so that the duration computation after the
  command runs can exclude time spent running plugin before_command
  listeners

---

For posterity, this also represents a soft decision about whether to
treat the `swain` variable (and the work to update it) as an optional
part of shellswain's API. It isn't essential for shellswain to provide
its profile-orchestration events, so it could be seen as pure overhead
by any modules that don't use it.

Since shellswain exists to help plugins avoid competing for resources
and duplicating work, I think it makes sense to provide a core set of
details that plugins are likely to compute/track on their own.